new attempt to refactor FRI
